+++ docs/ChangeLog.md
@@ -17,6 +17,7 @@
 
 * `$this->request()` returns not null but error message when `show_404()` or `show_error()` is called.
 * `$this->CI` in `TestCase` class after calling `$this->request()` is still an old instance.
+* Autoloader only searches class files only in top level and sub folder for them.
 
 ### Removed